Added the BFGamesManager
Modified the BFFriendsListController to show more extended friend information, the way you expect blackfire to handle it.

+++ BlackFire/BFGamesManager.h
@@ -13,5 +13,5 @@
 	
 }
 
-- (NSImage *)imageForGame:(unsigned int)gameID;
+- (NSImage *)imageForGame:(NSUInteger)gameID;
 @end
